move tests never running in jenkins to new evil tests module
