+++ core/src/test/java/org/elasticsearch/bootstrap/ESPolicyTests.java
@@ -21,57 +21,17 @@ package org.elasticsearch.bootstrap;
 
 import org.elasticsearch.test.ESTestCase;
 
-import java.io.FilePermission;
 import java.security.AccessControlContext;
 import java.security.AccessController;
-import java.security.AllPermission;
-import java.security.CodeSource;
-import java.security.Permission;
 import java.security.PermissionCollection;
 import java.security.Permissions;
 import java.security.PrivilegedAction;
 import java.security.ProtectionDomain;
-import java.security.cert.Certificate;
-import java.util.Collections;
 
 /** 
  * Tests for ESPolicy
- * <p>
- * Most unit tests won't run under security manager, since we don't allow 
- * access to the policy (you cannot construct it)
  */
 public class ESPolicyTests extends ESTestCase {
-
-    /** 
-     * Test policy with null codesource.
-     * <p>
-     * This can happen when restricting privileges with doPrivileged,
-     * even though ProtectionDomain's ctor javadocs might make you think
-     * that the policy won't be consulted.
-     */
-    public void testNullCodeSource() throws Exception {
-        assumeTrue("test cannot run with security manager", System.getSecurityManager() == null);
-        // create a policy with AllPermission
-        Permission all = new AllPermission();
-        PermissionCollection allCollection = all.newPermissionCollection();
-        allCollection.add(all);
-        ESPolicy policy = new ESPolicy(allCollection, Collections.emptyMap());
-        // restrict ourselves to NoPermission
-        PermissionCollection noPermissions = new Permissions();
-        assertFalse(policy.implies(new ProtectionDomain(null, noPermissions), new FilePermission("foo", "read")));
-    }
-
-    /** 
-     * test with null location
-     * <p>
-     * its unclear when/if this happens, see https://bugs.openjdk.java.net/browse/JDK-8129972
-     */
-    public void testNullLocation() throws Exception {
-        assumeTrue("test cannot run with security manager", System.getSecurityManager() == null);
-        PermissionCollection noPermissions = new Permissions();
-        ESPolicy policy = new ESPolicy(noPermissions, Collections.emptyMap());
-        assertFalse(policy.implies(new ProtectionDomain(new CodeSource(null, (Certificate[])null), noPermissions), new FilePermission("foo", "read")));
-    }
 
     /** 
      * test restricting privileges to no permissions actually works
