+++ qa/evil-tests/src/test/java/org/elasticsearch/plugins/PluginSecurityTests.java
