+++ qa/evil-tests/src/test/resources/org/elasticsearch/plugins/security/complex-plugin-security.policy
