+++ qa/evil-tests/src/test/resources/org/elasticsearch/plugins/security/simple-plugin-security.policy
