+++ qa/evil-tests/src/test/resources/org/elasticsearch/plugins/security/unresolved-plugin-security.policy
